+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Fix hanging publishing process
+
 ## [3.7.3] - 2026-07-29
 
 ### Added
@@ -290,6 +296,7 @@ confirmation prompt).
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_publish/compare/3.7.3...HEAD
 [3.7.3]: https://github.com/inlavigo/gg_publish/compare/3.7.2...3.7.3
 [3.7.2]: https://github.com/inlavigo/gg_publish/compare/3.7.1...3.7.2
 [3.7.1]: https://github.com/inlavigo/gg_publish/compare/3.7.0...3.7.1